// web/src/gen/models/TriageResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const dataStatusEnum = {
    PENDENTE: "PENDENTE",
    EM_ANALISE: "EM_ANALISE",
    CANCELADO: "CANCELADO",
    REENCAMINHADO: "REENCAMINHADO"
} as const;

export type DataStatusEnumKey = (typeof dataStatusEnum)[keyof typeof dataStatusEnum];

export const dataPriorityEnum = {
    BAIXA: "BAIXA",
    MEDIA: "MEDIA",
    ALTA: "ALTA"
} as const;

export type DataPriorityEnumKey = (typeof dataPriorityEnum)[keyof typeof dataPriorityEnum];

export type TriageResponse = {
    /**
     * @type object | undefined
    */
    data?: {
        /**
         * @type string | undefined
        */
        id?: string;
        /**
         * @type string | undefined
        */
        complaint?: string;
        /**
         * @type string | undefined
        */
        symptomDuration?: string;
        /**
         * @type string | undefined
        */
        symptom?: string;
        /**
         * @type string | undefined
        */
        actionTaken?: string;
        /**
         * @type string | undefined
        */
        reactionAfterAction?: string;
        /**
         * @type string | undefined
        */
        status?: DataStatusEnumKey;
        /**
         * @type string | undefined
        */
        patientId?: string;
        /**
         * @type string | undefined
        */
        analyzedById?: string;
        /**
         * @type string | undefined
        */
        specialtyId?: string;
        /**
         * @type string | undefined
        */
        consultationTypeId?: string;
        /**
         * @type string | undefined
        */
        priority?: DataPriorityEnumKey;
        /**
         * @type string | undefined
        */
        rejectionReason?: string;
        /**
         * @type string | undefined, date-time
        */
        createdAt?: string;
        /**
         * @type string | undefined, date-time
        */
        updatedAt?: string;
        /**
         * @type object | undefined
        */
        patient?: {
            /**
             * @type string | undefined
            */
            id?: string;
            /**
             * @type string | undefined
            */
            firstName?: string;
            /**
             * @type string | undefined
            */
            lastName?: string;
            /**
             * @type string | undefined
            */
            email?: string;
            /**
             * @type string | undefined
            */
            phone?: string;
        };
        /**
         * @type object | undefined
        */
        analyzedBy?: {
            /**
             * @type string | undefined
            */
            id?: string;
            /**
             * @type string | undefined
            */
            firstName?: string;
            /**
             * @type string | undefined
            */
            lastName?: string;
        };
        /**
         * @type object | undefined
        */
        specialty?: {
            /**
             * @type string | undefined
            */
            id?: string;
            /**
             * @type string | undefined
            */
            name?: string;
        };
        /**
         * @type object | undefined
        */
        consultationType?: {
            /**
             * @type string | undefined
            */
            id?: string;
            /**
             * @type string | undefined
            */
            name?: string;
        };
    };
};